[RSDK-8124] Move serve_web_with_external_network to common
