+++ micro-rdk/src/esp32/entry.rs
@@ -10,7 +10,7 @@ use crate::{
         conn::network::Network,
         credentials_storage::{RobotConfigurationStorage, WifiCredentialStorage},
         entry::RobotRepresentation,
-        entry::{serve_async_with_external_network, validate_robot_credentials},
+        entry::validate_robot_credentials,
         exec::Executor,
         grpc::ServerError,
         grpc_client::GrpcClientError,
@@ -249,56 +249,6 @@ pub fn serve_web<S>(
         info,
         storage,
         repr,
-        max_webrtc_connection,
-    )));
-
-    unreachable!()
-}
-
-pub fn serve_web_with_external_network<S>(
-    info: Option<ProvisioningInfo>,
-    repr: RobotRepresentation,
-    max_webrtc_connection: usize,
-    storage: S,
-    network: impl Network,
-) where
-    S: RobotConfigurationStorage + WifiCredentialStorage + Clone + 'static,
-    <S as RobotConfigurationStorage>::Error: Debug,
-    ServerError: From<<S as RobotConfigurationStorage>::Error>,
-    <S as WifiCredentialStorage>::Error: Sync + Send + 'static,
-{
-    // set the TWDT to expire after 5 minutes
-    crate::esp32::esp_idf_svc::sys::esp!(unsafe {
-        crate::esp32::esp_idf_svc::sys::esp_task_wdt_init(300, true)
-    })
-    .unwrap();
-
-    // Register the current task on the TWDT. The TWDT runs in the IDLE Task.
-    crate::esp32::esp_idf_svc::sys::esp!(unsafe {
-        crate::esp32::esp_idf_svc::sys::esp_task_wdt_add(
-            crate::esp32::esp_idf_svc::sys::xTaskGetCurrentTaskHandle(),
-        )
-    })
-    .unwrap();
-
-    let exec = Executor::new();
-    let cloned_exec = exec.clone();
-
-    cloned_exec
-        .spawn(async {
-            loop {
-                Timer::after(Duration::from_secs(150)).await;
-                unsafe { crate::esp32::esp_idf_svc::sys::esp_task_wdt_reset() };
-            }
-        })
-        .detach();
-
-    let _ = cloned_exec.block_on(Box::pin(serve_async_with_external_network(
-        exec,
-        info,
-        storage,
-        repr,
-        network,
         max_webrtc_connection,
     )));
