+++ micro-rdk/src/native/entry.rs
@@ -1,41 +1,3 @@
-#![allow(dead_code)]
-
-use std::fmt::Debug;
-
-use crate::common::{
-    conn::network::Network,
-    credentials_storage::{RobotConfigurationStorage, WifiCredentialStorage},
-    entry::{serve_async_with_external_network, RobotRepresentation},
-    exec::Executor,
-    grpc::ServerError,
-    provisioning::server::ProvisioningInfo,
-};
-
-pub fn serve_web_with_external_network<S>(
-    info: Option<ProvisioningInfo>,
-    repr: RobotRepresentation,
-    max_webrtc_connection: usize,
-    storage: S,
-    network: impl Network,
-) where
-    S: RobotConfigurationStorage + WifiCredentialStorage + Clone + 'static,
-    <S as RobotConfigurationStorage>::Error: Debug,
-    ServerError: From<<S as RobotConfigurationStorage>::Error>,
-    <S as WifiCredentialStorage>::Error: Sync + Send + 'static,
-{
-    let exec = Executor::new();
-    let cloned_exec = exec.clone();
-
-    let _ = cloned_exec.block_on(Box::pin(serve_async_with_external_network(
-        exec,
-        info,
-        storage,
-        repr,
-        network,
-        max_webrtc_connection,
-    )));
-}
-
 #[cfg(test)]
 mod tests {
     use crate::common::app_client::AppClientBuilder;
